backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
anthropic==0.28.0
langgraph==0.1.5
langchain-anthropic==0.1.15
langchain-core==0.2.5
gdeltdoc==1.4.0
earthengine-api==0.1.402
pydantic==2.7.1
